chore(maths): missing tests for add
A new function was introduced in b8f7566 but the unittest were
missings. This commit adds a way to test the code along with some tests
for that function.

To run the tests simply start `maths.py` right away instead of importing
it. It exists with a non-zero status code and a message when a test
break. It prints "ok" and exits with a zero status code when all tests
passed.

    $ python3 maths.py  # "py maths.py" on windows
    ok
    $