chore(maths): sort functions by name
The current order of functions follow a sense of "complexity", i.e. the
addition is before the substraction and before the multiplication
because it is an "easier" operation. As there are more and more
functions added into the module, it becomes clear that it would be
better to sort them by name to ease navigating into the source code.