chore(maths): use doctest to test docstring example
The doctest[^1] module is a special library available in the standard
library of python that can be used to evaluate example snippets inside
of function docstrings. Because most of our tests are very simple and
would be great examples of how to use the functions they test, we
decided to change `assert` for `doctest`.

In the following example:

    def foo():
        """
        Compute the square root of 25.

        >>> foo()
        5
        """
        return 5

Doctest is going to execute `foo()` and it verifies that the function
returns `5`. Any text that precedes the first example (noted by `>>>`)
is ignored.

[^1]: https://pymotw.com/3/doctest/